Extract dataframe methods
