+++ src/core/dataframe.js
@@ -1,3 +1,5 @@
+import * as earcut from 'earcut';
+
 /**
  * @typedef {object} Dataframe - Point in renderer coordinates space
  * @property {RPoint} center
@@ -5,13 +7,107 @@
  * @property {geom} geometry
  * @property {Properties} properties
  */
-
 export default class Dataframe {
     constructor(center, scale, geom, properties) {
         this.center = center;
         this.scale = scale;
         this.geom = geom;
         this.properties = properties;
+    }
+    bind(renderer) {
+        const gl = renderer.gl;
+        this.renderer = renderer;
+
+        this.propertyTex = [];
+
+        const decodedGeom = decodeGeom(this.type, this.geom);
+        var points = decodedGeom.geometry;
+        this.numVertex = points.length / 2;
+        this.breakpointList = decodedGeom.breakpointList;
+        this.numFeatures = this.breakpointList.length || this.numVertex;
+        this._genDataframePropertyTextures(gl);
+
+        const width = this.renderer.RTT_WIDTH;
+        const height = Math.ceil(this.numFeatures / width);
+
+        this.setStyle = (style) => {
+            this.style = style;
+        };
+        this.style = null;
+
+        this.vertexBuffer = gl.createBuffer();
+        this.featureIDBuffer = gl.createBuffer();
+
+        this.texColor = this._createStyleTileTexture(this.numFeatures);
+        this.texWidth = this._createStyleTileTexture(this.numFeatures);
+        this.texStrokeColor = this._createStyleTileTexture(this.numFeatures);
+        this.texStrokeWidth = this._createStyleTileTexture(this.numFeatures);
+
+        var ids = new Float32Array(points.length);
+        let index = 0;
+        for (var i = 0; i < points.length; i += 2) {
+            if ((!this.breakpointList.length && i > 0) || i == this.breakpointList[index]) {
+                index++;
+            }
+            ids[i + 0] = ((index) % width) / (width - 1);
+            ids[i + 1] = height > 1 ? Math.floor((index) / width) / (height - 1) : 0.5;
+        }
+        gl.bindBuffer(gl.ARRAY_BUFFER, this.vertexBuffer);
+        gl.bufferData(gl.ARRAY_BUFFER, points, gl.STATIC_DRAW);
+
+        if (decodedGeom.normals) {
+            this.normalBuffer = gl.createBuffer();
+            gl.bindBuffer(gl.ARRAY_BUFFER, this.normalBuffer);
+            gl.bufferData(gl.ARRAY_BUFFER, decodedGeom.normals, gl.STATIC_DRAW);
+        }
+
+        gl.bindBuffer(gl.ARRAY_BUFFER, this.featureIDBuffer);
+        gl.bufferData(gl.ARRAY_BUFFER, ids, gl.STATIC_DRAW);
+    }
+    _genDataframePropertyTextures() {
+        const gl = this.renderer.gl;
+        const width = this.renderer.RTT_WIDTH;
+        const height = Math.ceil(this.numFeatures / width);
+
+        this.height = height;
+        this.propertyID = {}; //Name => PID
+        this.propertyCount = 0;
+        for (var k in this.properties) {
+            if (this.properties.hasOwnProperty(k) && this.properties[k].length > 0) {
+                var propertyID = this.propertyID[k];
+                if (propertyID === undefined) {
+                    propertyID = this.propertyCount;
+                    this.propertyCount++;
+                    this.propertyID[k] = propertyID;
+                }
+                this.propertyTex[propertyID] = gl.createTexture();
+                gl.bindTexture(gl.TEXTURE_2D, this.propertyTex[propertyID]);
+                gl.texImage2D(gl.TEXTURE_2D, 0, gl.ALPHA,
+                    width, height, 0, gl.ALPHA, gl.FLOAT,
+                    this.properties[k]);
+                gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
+                gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
+                gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST);
+                gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST);
+            }
+        }
+    }
+    _createStyleTileTexture(numFeatures) {
+        // TODO we are wasting 75% of the memory for the scalar attributes (width, strokeWidth),
+        // since RGB components are discarded
+        const gl = this.renderer.gl;
+        const width = this.renderer.RTT_WIDTH;
+        const height = Math.ceil(numFeatures / width);
+        const texture = gl.createTexture();
+        gl.bindTexture(gl.TEXTURE_2D, texture);
+        gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA,
+            width, height, 0, gl.RGBA, gl.UNSIGNED_BYTE,
+            null);
+        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
+        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
+        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST);
+        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST);
+        return texture;
     }
     free() {
         if (this.propertyTex) {
@@ -30,5 +126,134 @@ export default class Dataframe {
             this.featureIDBuffer = 'freed';
             this.propertyTex = null;
         }
+    }
+}
+
+
+function getLineNormal(a, b) {
+    const dx = b[0] - a[0];
+    const dy = b[1] - a[1];
+    return normalize([-dy, dx]);
+}
+
+function normalize(v) {
+    const s = Math.sqrt(v[0] * v[0] + v[1] * v[1]);
+    return [v[0] / s, v[1] / s];
+}
+// Decode a tile geometry
+// If the geometry type is 'point' it will pass trough the geom (the vertex array)
+// If the geometry type is 'polygon' it will triangulate the polygon list (geom)
+//      geom will be a list of polygons in which each polygon will have a flat array of vertices and a list of holes indices
+//      Example:
+/*         let geom = [
+                {
+                    flat: [
+                        0.,0., 1.,0., 1.,1., 0.,1., 0.,0, //A square
+                        0.25,0.25, 0.75,0.25, 0.75,0.75, 0.25,0.75, 0.25,0.25//A small square
+                    ]
+                    holes: [5]
+                }
+            ]
+*/
+// If the geometry type is 'line' it will generate the appropriate zero-sized, vertex-shader expanded triangle list with mitter joints.
+// The geom will be an array of coordinates in this case
+function decodeGeom(geomType, geom) {
+    if (geomType == 'point') {
+        return {
+            geometry: geom,
+            breakpointList: []
+        };
+    } else if (geomType == 'polygon') {
+        let vertexArray = []; //Array of triangle vertices
+        let breakpointList = []; // Array of indices (to vertexArray) that separate each feature
+        geom.map(feature => {
+            feature.map(polygon => {
+                const triangles = earcut(polygon.flat, polygon.holes);
+                triangles.map(index => {
+                    vertexArray.push(polygon.flat[2 * index]);
+                    vertexArray.push(polygon.flat[2 * index + 1]);
+                });
+            });
+            breakpointList.push(vertexArray.length);
+        });
+        return {
+            geometry: new Float32Array(vertexArray),
+            breakpointList
+        };
+    } else if (geomType == 'line') {
+        let geometry = [];
+        let normals = [];
+        let breakpointList = []; // Array of indices (to vertexArray) that separate each feature
+        geom.map(feature => {
+            feature.map(line => {
+                // Create triangulation
+                for (let i = 0; i < line.length - 2; i += 2) {
+                    const a = [line[i + 0], line[i + 1]];
+                    const b = [line[i + 2], line[i + 3]];
+                    if (i > 0) {
+                        var prev = [line[i + -2], line[i + -1]];
+                        var nprev = getLineNormal(a, prev);
+                    }
+                    if (i < line.length - 4) {
+                        var next = [line[i + 4], line[i + 5]];
+                        var nnext = getLineNormal(next, b);
+                    }
+                    let normal = getLineNormal(b, a);
+                    let na = normal;
+                    let nb = normal;
+                    //TODO bug, cartesian interpolation is not correct, should use polar coordinates for the interpolation
+                    if (prev) {
+                        na = normalize([
+                            normal[0] * 0.5 + nprev[0] * 0.5,
+                            normal[1] * 0.5 + nprev[1] * 0.5,
+                        ]);
+                    }
+                    if (next) {
+                        nb = normalize([
+                            normal[0] * 0.5 + nnext[0] * 0.5,
+                            normal[1] * 0.5 + nnext[1] * 0.5,
+                        ]);
+                    }
+                    normals.push(-na[0], -na[1]);
+                    normals.push(na[0], na[1]);
+                    normals.push(-nb[0], -nb[1]);
+
+                    normals.push(na[0], na[1]);
+                    normals.push(nb[0], nb[1]);
+                    normals.push(-nb[0], -nb[1]);
+
+                    normal = [0, 0];
+
+
+                    //First triangle
+                    geometry.push(a[0] - 0.01 * normal[0]);
+                    geometry.push(a[1] - 0.01 * normal[1]);
+
+                    geometry.push(a[0] + 0.01 * normal[0]);
+                    geometry.push(a[1] + 0.01 * normal[1]);
+
+                    geometry.push(b[0] - 0.01 * normal[0]);
+                    geometry.push(b[1] - 0.01 * normal[1]);
+
+                    //Second triangle
+                    geometry.push(a[0] + 0.01 * normal[0]);
+                    geometry.push(a[1] + 0.01 * normal[1]);
+
+                    geometry.push(b[0] + 0.01 * normal[0]);
+                    geometry.push(b[1] + 0.01 * normal[1]);
+
+                    geometry.push(b[0] - 0.01 * normal[0]);
+                    geometry.push(b[1] - 0.01 * normal[1]);
+                }
+            });
+            breakpointList.push(geometry.length);
+        });
+        return {
+            geometry: new Float32Array(geometry),
+            breakpointList,
+            normals: new Float32Array(normals)
+        };
+    } else {
+        throw new Error(`Unimplemented geometry type: '${geomType}'`);
     }
 }
