+++ src/core/renderer.js
@@ -1,6 +1,5 @@
 import * as shaders from './shaders';
 import * as schema from './schema';
-import * as earcut from 'earcut';
 import Dataframe from './dataframe';
 import { Asc, Desc } from './style/functions';
 
@@ -46,6 +45,7 @@ class Renderer {
         }
         this._center = { x: 0, y: 0 };
         this._zoom = 1;
+        this.RTT_WIDTH = RTT_WIDTH;
         console.log('R', this);
         this.dataframes = [];
         this.computePool = []; //TODO hack, refactor needed
@@ -149,54 +149,6 @@ class Renderer {
         this.dataframes = this.dataframes.filter(t => t !== dataframe);
     }
 
-    createStyleTileTexture(numFeatures) {
-        // TODO we are wasting 75% of the memory for the scalar attributes (width, strokeWidth),
-        // since RGB components are discarded
-        const gl = this.gl;
-        const width = RTT_WIDTH;
-        const height = Math.ceil(numFeatures / width);
-        const texture = gl.createTexture();
-        gl.bindTexture(gl.TEXTURE_2D, texture);
-        gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA,
-            width, height, 0, gl.RGBA, gl.UNSIGNED_BYTE,
-            null);
-        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
-        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
-        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST);
-        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST);
-        return texture;
-    }
-
-    _genDataframePropertyTextures(dataframe){
-        const gl = this.gl;
-        const width = RTT_WIDTH;
-        const height = Math.ceil(dataframe.numFeatures / width);
-
-        dataframe.height = height;
-        dataframe.propertyID = {}; //Name => PID
-        dataframe.propertyCount = 0;
-        dataframe.renderer = this;
-        for (var k in dataframe.properties) {
-            if (dataframe.properties.hasOwnProperty(k) && dataframe.properties[k].length > 0) {
-                var propertyID = dataframe.propertyID[k];
-                if (propertyID === undefined) {
-                    propertyID = dataframe.propertyCount;
-                    dataframe.propertyCount++;
-                    dataframe.propertyID[k] = propertyID;
-                }
-                dataframe.propertyTex[propertyID] = gl.createTexture();
-                gl.bindTexture(gl.TEXTURE_2D, dataframe.propertyTex[propertyID]);
-                gl.texImage2D(gl.TEXTURE_2D, 0, gl.ALPHA,
-                    width, height, 0, gl.ALPHA, gl.FLOAT,
-                    dataframe.properties[k]);
-                gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
-                gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
-                gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST);
-                gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST);
-            }
-        }
-    }
-
     /**
      * @description Adds a new dataframe to the renderer.
      *
@@ -206,56 +158,8 @@ class Renderer {
      * @returns {BoundDataframe}
      */
     addDataframe(dataframe) {
-        const gl = this.gl;
-        //this.ext.bindVertexArrayOES(this.vao);
+        dataframe.bind(this);
         this.dataframes.push(dataframe);
-        dataframe.propertyTex = [];
-
-        const decodedGeom = decodeGeom(dataframe.type, dataframe.geom);
-        var points = decodedGeom.geometry;
-        dataframe.numVertex = points.length / 2;
-        dataframe.breakpointList = decodedGeom.breakpointList;
-        dataframe.numFeatures = dataframe.breakpointList.length || dataframe.numVertex;
-        this._genDataframePropertyTextures(dataframe);
-
-        const width = RTT_WIDTH;
-        const height = Math.ceil(dataframe.numFeatures / width);
-
-        dataframe.setStyle = (style) => {
-            dataframe.style = style;
-        };
-        dataframe.style = null;
-
-        dataframe.vertexBuffer = gl.createBuffer();
-        dataframe.featureIDBuffer = gl.createBuffer();
-
-        dataframe.texColor = this.createStyleTileTexture(dataframe.numFeatures);
-        dataframe.texWidth = this.createStyleTileTexture(dataframe.numFeatures);
-        dataframe.texStrokeColor = this.createStyleTileTexture(dataframe.numFeatures);
-        dataframe.texStrokeWidth = this.createStyleTileTexture(dataframe.numFeatures);
-
-        var ids = new Float32Array(points.length);
-        let index = 0;
-        for (var i = 0; i < points.length; i += 2) {
-            if ((!dataframe.breakpointList.length && i > 0) || i == dataframe.breakpointList[index]) {
-                index++;
-            }
-            ids[i + 0] = ((index) % width) / (width - 1);
-            ids[i + 1] = height > 1 ? Math.floor((index) / width) / (height - 1) : 0.5;
-        }
-        gl.bindBuffer(gl.ARRAY_BUFFER, dataframe.vertexBuffer);
-        gl.bufferData(gl.ARRAY_BUFFER, points, gl.STATIC_DRAW);
-
-        if (decodedGeom.normals) {
-            dataframe.normalBuffer = gl.createBuffer();
-            gl.bindBuffer(gl.ARRAY_BUFFER, dataframe.normalBuffer);
-            gl.bufferData(gl.ARRAY_BUFFER, decodedGeom.normals, gl.STATIC_DRAW);
-        }
-
-        gl.bindBuffer(gl.ARRAY_BUFFER, dataframe.featureIDBuffer);
-        gl.bufferData(gl.ARRAY_BUFFER, ids, gl.STATIC_DRAW);
-
-        return dataframe;
     }
 
     getAspect() {
@@ -584,137 +488,6 @@ class Renderer {
     }
 
 }
-
-
-function getLineNormal(a, b) {
-    const dx = b[0] - a[0];
-    const dy = b[1] - a[1];
-    return normalize([-dy, dx]);
-}
-
-function normalize(v) {
-    const s = Math.sqrt(v[0] * v[0] + v[1] * v[1]);
-    return [v[0] / s, v[1] / s];
-}
-// Decode a tile geometry
-// If the geometry type is 'point' it will pass trough the geom (the vertex array)
-// If the geometry type is 'polygon' it will triangulate the polygon list (geom)
-//      geom will be a list of polygons in which each polygon will have a flat array of vertices and a list of holes indices
-//      Example:
-/*         let geom = [
-                {
-                    flat: [
-                        0.,0., 1.,0., 1.,1., 0.,1., 0.,0, //A square
-                        0.25,0.25, 0.75,0.25, 0.75,0.75, 0.25,0.75, 0.25,0.25//A small square
-                    ]
-                    holes: [5]
-                }
-            ]
-*/
-// If the geometry type is 'line' it will generate the appropriate zero-sized, vertex-shader expanded triangle list with mitter joints.
-// The geom will be an array of coordinates in this case
-function decodeGeom(geomType, geom) {
-    if (geomType == 'point') {
-        return {
-            geometry: geom,
-            breakpointList: []
-        };
-    } else if (geomType == 'polygon') {
-        let vertexArray = []; //Array of triangle vertices
-        let breakpointList = []; // Array of indices (to vertexArray) that separate each feature
-        geom.map(feature => {
-            feature.map(polygon => {
-                const triangles = earcut(polygon.flat, polygon.holes);
-                triangles.map(index => {
-                    vertexArray.push(polygon.flat[2 * index]);
-                    vertexArray.push(polygon.flat[2 * index + 1]);
-                });
-            });
-            breakpointList.push(vertexArray.length);
-        });
-        return {
-            geometry: new Float32Array(vertexArray),
-            breakpointList
-        };
-    } else if (geomType == 'line') {
-        let geometry = [];
-        let normals = [];
-        let breakpointList = []; // Array of indices (to vertexArray) that separate each feature
-        geom.map(feature => {
-            feature.map(line => {
-                // Create triangulation
-                for (let i = 0; i < line.length - 2; i += 2) {
-                    const a = [line[i + 0], line[i + 1]];
-                    const b = [line[i + 2], line[i + 3]];
-                    if (i > 0) {
-                        var prev = [line[i + -2], line[i + -1]];
-                        var nprev = getLineNormal(a, prev);
-                    }
-                    if (i < line.length - 4) {
-                        var next = [line[i + 4], line[i + 5]];
-                        var nnext = getLineNormal(next, b);
-                    }
-                    let normal = getLineNormal(b, a);
-                    let na = normal;
-                    let nb = normal;
-                    //TODO bug, cartesian interpolation is not correct, should use polar coordinates for the interpolation
-                    if (prev) {
-                        na = normalize([
-                            normal[0] * 0.5 + nprev[0] * 0.5,
-                            normal[1] * 0.5 + nprev[1] * 0.5,
-                        ]);
-                    }
-                    if (next) {
-                        nb = normalize([
-                            normal[0] * 0.5 + nnext[0] * 0.5,
-                            normal[1] * 0.5 + nnext[1] * 0.5,
-                        ]);
-                    }
-                    normals.push(-na[0], -na[1]);
-                    normals.push(na[0], na[1]);
-                    normals.push(-nb[0], -nb[1]);
-
-                    normals.push(na[0], na[1]);
-                    normals.push(nb[0], nb[1]);
-                    normals.push(-nb[0], -nb[1]);
-
-                    normal = [0, 0];
-
-
-                    //First triangle
-                    geometry.push(a[0] - 0.01 * normal[0]);
-                    geometry.push(a[1] - 0.01 * normal[1]);
-
-                    geometry.push(a[0] + 0.01 * normal[0]);
-                    geometry.push(a[1] + 0.01 * normal[1]);
-
-                    geometry.push(b[0] - 0.01 * normal[0]);
-                    geometry.push(b[1] - 0.01 * normal[1]);
-
-                    //Second triangle
-                    geometry.push(a[0] + 0.01 * normal[0]);
-                    geometry.push(a[1] + 0.01 * normal[1]);
-
-                    geometry.push(b[0] + 0.01 * normal[0]);
-                    geometry.push(b[1] + 0.01 * normal[1]);
-
-                    geometry.push(b[0] - 0.01 * normal[0]);
-                    geometry.push(b[1] - 0.01 * normal[1]);
-                }
-            });
-            breakpointList.push(geometry.length);
-        });
-        return {
-            geometry: new Float32Array(geometry),
-            breakpointList,
-            normals: new Float32Array(normals)
-        };
-    } else {
-        throw new Error(`Unimplemented geometry type: '${geomType}'`);
-    }
-}
-
-
 
 class ComputeJob {
     constructor(type, expressions, resolve) {
